+++ src/main/java/com/pepe/shoppingcart/repository/UsersRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.pepe.shoppingcart.model.Products;
 import com.pepe.shoppingcart.model.Users;
 
-public interface UsersRepository extends JpaRepository<Users, Integer> {
+public interface UsersRepository extends JpaRepository<Users, Long> {
 	Optional<Users> findByEmail(String email);
 	List<Users> findByUserName(String userName);
 }
